+++ src/features/account/controllers/account-controller.ts
@@ -1,25 +1,25 @@
-import { accountDeleteBySchema } from '@/features/account/validators/account-find-by-id-schema.js';
-import type { Validator } from '@/shared/infra/validator/validator.js';
-import type { Controller } from '@/shared/protocols/controller.js';
-import type { AsyncRequestHandler } from '@/shared/protocols/handlers.js';
-import { HttpStatusCode } from '@/shared/protocols/http-client.js';
-import type { Service } from '@/shared/protocols/service.js';
+import type { Controller } from '@/shared/protocols/controller';
+import type { AsyncRequestHandler } from '@/shared/protocols/handlers';
+import { HttpStatusCode } from '@/shared/protocols/http-client';
+
+import type { DeleteUserAccountsService } from '../services/delete-user-accounts-service';
 
 export class AccountController implements Controller {
-  deleteAccountById: AsyncRequestHandler = async (req, res) => {
-    const accountId = req.params.id;
+  deleteAccountById: AsyncRequestHandler = async (req, res, next) => {
+    try {
+      const accountId = Number(req.params.id);
 
-    this.validator.validate(accountDeleteBySchema, {
-      params: req.params,
-    });
+      await this.deleteUserAccountService.execute({
+        socialMediaId: accountId,
+      });
 
-    await this.deleteByIdService.execute({ socialMediaId: accountId });
-
-    res.status(HttpStatusCode.noContent).send();
+      return res.status(HttpStatusCode.noContent).send();
+    } catch (error) {
+      next(error);
+    }
   };
 
   constructor(
-    private readonly validator: Validator,
-    private readonly deleteByIdService: Service
+    private readonly deleteUserAccountService: DeleteUserAccountsService
   ) {}
 }
